Implement grid system using flex box and remove js folder
