+++ package.json
@@ -18,6 +18,7 @@
     "babel-preset-es2015": "^6.24.1",
     "css-loader": "^3.2.0",
     "file-loader": "^4.2.0",
+    "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.8.0",
     "node-sass": "^4.12.0",
