+++ src/components/details/details.js
