+++ src/components/details/details.scss
@@ -1,0 +1,9 @@
+app-details {
+  .image-container {
+    text-align: center;
+  }
+
+  h4 {
+    text-align: center;
+  }
+}
