+++ src/components/home/home.pug
@@ -1,7 +1,7 @@
 app-home
   .flex-container
     each character in characters
-      .flex-item
+      .col-xs-12.col-sm-4
         .card
           .card-image
             img(src=character.image)
