+++ src/components/home/home.scss
@@ -1,0 +1,5 @@
+app-home {
+  .btn {
+    margin-top: 10px;
+  }
+}
