+++ src/components/layout/footer.pug
@@ -1,0 +1,2 @@
+footer.page-footer.grey.darken-4
+  .footer-copyright.right © 2019 Copyright by Juan Diego Lopez
