+++ src/components/layout/main.pug
@@ -1,3 +1,3 @@
 include navbar.pug
-.content
+.content.container
 include footer.pug
