+++ src/components/layout/navbar.pug
@@ -1,6 +1,6 @@
 nav
-  .nav-wrapper.cyan.darken-2
-    a.brand-logo.right(href="#") Rick And Morty
+  .nav-wrapper.grey.darken-4
+    a.brand-logo.right(href="/")
     ul#nav-mobile.left.hide-on-med-and-down
       li
         a(href="/home") Home
