--- src/js/components/characters/characters.scss
@@ -1,25 +1,0 @@
-app-characters {
-  .flex-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-  }
-  
-  .flex-item {
-    flex: 20%;
-    padding: 0 4px;
-    margin-bottom: 10px;
-  
-    @media screen and (max-width: 1000px) {
-      flex: 50%;
-    }
-  
-    @media screen and (max-width: 600px) {
-      flex: 100%;
-    }
-  }
-  
-  .character {
-    cursor: pointer;
-  }
-}
