--- src/js/components/details/details.scss
@@ -1,21 +1,0 @@
-app-details {
-  .flex-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    padding: 0 4px;
-  }
-
-  .flex-item {
-    flex: 50%;
-    padding: 0 10px;
-
-    @media screen and (max-width: 600px) {
-      flex: 100%;
-    }
-  }
-
-  h4 {
-    text-align: center;
-  }
-}
