--- src/js/components/home/home.scss
@@ -1,23 +1,0 @@
-app-home {
-  .flex-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    padding: 0 4px;
-  }
-  
-  .flex-item {
-    flex: 25%;
-    max-width: 80%;
-    padding: 0 4px;
-    margin-bottom: 10px;
-  
-    @media screen and (max-width: 600px) {
-      flex: 100%;
-    }
-  }
-  
-  .btn {
-    margin-top: 10px;
-  }
-}
