--- src/js/components/layout/footer.pug
@@ -1,2 +1,0 @@
-footer.page-footer.cyan.darken-2
-  .footer-copyright © 2019 Copyright by Juan Diego Lopez
