+++ src/scss/style.scss
@@ -22,3 +22,27 @@
   height: 100%;
   margin: 0;
 }
+
+.brand-logo > img {
+  max-width: 180px;
+  
+  @media screen and (max-width: 600px) {
+    max-width: 150px;
+  }
+}
+
+#nav-mobile li:hover {
+  background-color: #424242;
+}
+
+.btn {
+  background-color: #00afc7;
+}
+
+.btn:hover {
+  background-color: #047d8d;
+}
+
+.page-footer .footer-copyright {
+  padding: 0px 20px;
+}
